Optimize fd_write
